requirements.txt
boto3==1.7.58
botocore==1.10.58
Django==2.0.7
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
Werkzeug==0.14.1
whitenoise==3.3.1
zappa-django-utils==0.4.0
